Intel issue DPD200416083
CAF wrapped into OOP

Intel Fortran Compiler seems to be confused when calling a TBP
associated to a caoarray component of a derived type. When the `add` TBP
of the test is invoked a SIGSEGV is generated.

This is currently an official issue for which deeper investigations are
pending, but not yet an official bug.

I reproduced the issue with Intel v. 16.0.3 and 17.0.1, under GNU Linux
disto (with Arch Linux and Ubuntu).

> to compile the test use `ifort -coarray oop-caf.f90`

Signed-off-by: Stefano Zaghi <stefano.zaghi@gmail.com>